Add a silly web extension
When run with DINGHY_WEB_EXTENSION_DEMO, dinghy will now print out the
page ID of every newly-created WebKitWebPage in the web process. This is
silly, but it shows that we can run code in the web process.

In the future, the demo functionality could be removed and replaced with
something that would actually be useful, e.g. an adblocker.