Add a WIP DyPlatform implementation for -fdo backend
For now it is almost a raw copy of wpe-fdo-view's implementation.
That means it expects a Wayland compositor running in the system,
no X11 support for now.

There is a lot of functionality in this platform that should be
moved to dinghy itself (e.g, key-bindings, toggling fullscreen,
resizing, etc).