dinghyctl: Do not segfault on unrecognized command names
When a non-existent command name was passed to cmd_find_by_name() the
program would crash due to attempting to compare the command name with
the NULL string from the terminating entry of the commands table.

This fixes the issue by making the loop do a NULL-check for termination
instead of iterating over the number of entries in the table (which
included the terminating entry).