Allow universal access from file:// URIs by default
This assumes that content from the local file system can be trusted
and allows it to perform requests (e.g. using XmlHttpRequest or the
Fetch API) which normally would be blocked due to the stricter checks
done on requests started from file:// resources.

It is possible to re-enable the stricter mode by passing the command
line option "--strict-file-uri-checking".