Make selected signals terminate the event loop gracefully
Install signal handlers for SIGHUP, SIGINT, SIGTERM, SIGUSR1, and
SIGUSR2 in order to stop the application event loop before exiting.

When one of these signals is received, its signal code is stored
and the main loop stopped with g_application_quit(). Once shutdown
completes and control returns from g_application_run(), the signal
which caused the main loop to end is re-raised in order to make the
process terminate in the way it was supposed to (i.e. terminated
by a signal instead of successful execution).

This should ensure correct finalization of child processes and
deallocation of resources before the process terminates, while
still maintaining the same externally observable behaviour.

Fixes #21